+++ app/assets/javascripts/notebook/cells/markdown.vue
@@ -157,6 +157,7 @@ export default {
           'var',
         ],
         ALLOWED_ATTR: ['class', 'style', 'href', 'src'],
+        ALLOW_DATA_ATTR: false,
       });
     },
   },